dev.requirements.txt
pytest==5.4.1
requests-mock==1.5.2
mkdocs==1.0.4
mkdocs-material==4.1.1
pytest-click==0.3
pytest-cov==2.8.1
coveralls==2.0.0
black==19.10b0
flake8==3.7.9
isort==4.3.21
